css/css-ui-3/resize-018.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Basic User Interface Test: resizing pictures</title>
<link rel="author" title="Florian Rivoal" href="http://florian.rivoal.net">
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#resize">
<meta name="flags" content="interact">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#test {
  position: absolute;
  width: 100px;
  height: 100px;
  background: orange;
  overflow: hidden;
  resize: both;
}
#target {
  position: absolute;
  width: 150px;
  height: 150px;
  background: blue;
}
#log { margin-top: 200px; }
</style>

<p>Enlarge the orange box so that it at least covers the blue box entirely, then press the “done” button.</p>
<!-- There's no actual need to cover the whole box,
  but doing so makes sure that the user has resized in both dimensions -->

<button onclick="verify()">done</button>

<div id="target"></div>
<div id="test"></div>
<div id=log></div>
<script>
setup({ explicit_done: true });
function verify() {
  test(
    function(){
      var test = document.getElementById("test");
      assert_regexp_match(test.style.width, /px$/, "The width property of the style attribute should be set");
      assert_regexp_match(test.style.height, /px$/, "The height property of the style attribute should be set");
      document.styleSheets[0].cssRules[0].style.resize= "none";
      assert_regexp_match(test.style.width, /px$/, "The width property of the style attribute should still be set after removing ability to resize");
      assert_regexp_match(test.style.height, /px$/, "The height property of the style attribute should still be set after removing ability to resize");
    }, "Removing the ability to resize an element does not reset its size");
  done();
}
</script>
